fix delay time in blocs, fix red button responsive size
