+++ lib/screens/signin/signin_screen.dart
@@ -6,8 +6,8 @@ import 'package:openflutterecommerce/screens/signin/signup.dart';
 import 'package:openflutterecommerce/screens/signin/views/right_arrow_action.dart';
 import 'package:openflutterecommerce/screens/signin/views/service_button.dart';
 import 'package:openflutterecommerce/screens/signin/views/sign_in_field.dart';
-import 'package:openflutterecommerce/screens/signin/views/signin_button.dart';
 import 'package:openflutterecommerce/screens/signin/views/title.dart';
+import 'package:openflutterecommerce/widgets/red_button_with_shadow.dart';
 
 import '../../config/routes.dart';
 import '../../config/theme.dart';
@@ -77,7 +77,7 @@ class _SignInScreenState extends State<SignInScreen> {
                   "Forgot your password",
                   onClick: _showForgotPassword,
                 ),
-                SignInButton("LOGIN", onPressed: _validateAndSend),
+                RedButtonWithShadow("LOGIN", onPressed: _validateAndSend),
                 SizedBox(
                   height: sizeBetween * 2,
                 ),
